Use size_t where appropriate (for pilot-link 0.12 this is a real issue)
svn path=/trunk/kdepim/; revision=310708

+++ kpilot/lib/pilotLocalDatabase.h
@@ -149,7 +149,7 @@ private:
 	struct DBInfo fDBInfo;
 	QString fPathName,fDBName;
 	char*       fAppInfo;
-	int         fAppLen;
+	size_t      fAppLen;
 	int         fNumRecords;
 	int         fCurrentRecord;
 	PilotRecord* fRecords[10000]; // Current max records in DB.. hope it's enough
